+++ Package.swift
@@ -14,7 +14,8 @@ let package = Package(
         .package(url: "https://github.com/swiftlang/swift-docc-plugin.git", from: "1.4.0"),
         .package(url: "https://github.com/bare-swift/swift-bytes.git", from: "0.1.0"),
         .package(url: "https://github.com/bare-swift/swift-varint.git", from: "0.1.0"),
-        .package(url: "https://github.com/bare-swift/swift-otlp-exporter.git", from: "0.1.0")
+        .package(url: "https://github.com/bare-swift/swift-otlp-exporter.git", from: "0.1.0"),
+        .package(url: "https://github.com/bare-swift/swift-time.git", from: "0.1.0")
     ],
     targets: [
         .target(
@@ -22,7 +23,8 @@ let package = Package(
             dependencies: [
                 .product(name: "Bytes", package: "swift-bytes"),
                 .product(name: "Varint", package: "swift-varint"),
-                .product(name: "OTLPExporter", package: "swift-otlp-exporter")
+                .product(name: "OTLPExporter", package: "swift-otlp-exporter"),
+                .product(name: "Time", package: "swift-time")
             ]
         ),
         .testTarget(